Keyless sharding for insert
Summary:
Pretty simple design.  Basically if you set the isKeylessSharded flag, we call show partitions and pick a random partition (colocated if possible).  We make no effort to verify your table is actually keyless.
For pinterest, this was a 5x perf win (but it isn't apples to apples because we didnt have colocation), at the expense of noticably higher shard variance.

Test Plan: basic + litterbug

Reviewers: hurshal, ankur, wsong, carl

Subscribers: engineering-list

Differential Revision: http://grizzly.memsql.com/D11628