Don't compress colocated loads.
Summary: Depending on the number of partitions, Spark can spend an aweful lot of time in compression.  With a heavily coalesced workload (as was the case before the keyless optimization), not compressing is shown to decrease batch latency significantly.  There is (maybe) no risk of blowing out a nodes network and causing heartbeat failure in the colocated case, so no need to compress.

Test Plan: litterbug, existing

Reviewers: ankur, carl, wsong, hurshal

Subscribers: engineering-list

Differential Revision: http://grizzly.memsql.com/D11622