Fix stopping pipelines by ensuring we only have one PipelineMonitor per pipeline
Summary:
The issue was that we were using .map on a future instead of getting the result of the future in the main loop of the Superapp. If creating a PipelineMonitor took longer than 5 seconds, this will result in multiple PipelineMonitor threads being created for a single pipeline, but we would only stop one when we marked a pipeline as stopped. This code blocks on the PipelineMonitor creation code, so we don't wind up with duplicate pipeline monitors.

It also appears that thread.interrupt does not stop the thread as we hoped, so I used an AtomicBoolean to stop PipelineMonitors.

Test Plan: manual

Reviewers: ankur, joseph, hurshal, carl

Reviewed By: carl

Subscribers: engineering-list

Differential Revision: http://grizzly.memsql.com/D11760