MemSQLExtraColumn
Summary:
This allows you to specify extra columns with creating a table from a dataframe.
As discussed, this is reasonably general, but the interface can expose whatever restricted thing it wants to the user.
The following LoadPhaseConfig should allow you to create a columnstore table sorted on insert_time
{
  "table_extra_columns" :
   [{ "name": "insert_time",
     "type": "TIMESTAMP",
     "default_value": "CURRENT_TIME"
   }]
  "table_keys" :
  [{
    "key_type" : "KeyUsingClusteredColumnStore",
    "columns":["insert_time"]
  }]
}

Test Plan: ..

Reviewers: wsong, carl

Subscribers: engineering-list

Differential Revision: http://grizzly.memsql.com/D12043