Set debug logging in Akka logging as well when --debug is specified in the interface
Summary: This will ensure that HTTP API requests show up when --debug is specified. Note that Akka debug logging is pretty spammy (i.e. it includes a lot more than just our HTTP API requests), so if this turns out to be too much, we can log requests and responses at the INFO level and set the default Akka level to WARNING or something like that.

Test Plan: manual

Reviewers: carl, ankur, joseph, hurshal

Reviewed By: hurshal

Subscribers: engineering-list

Differential Revision: http://grizzly.memsql.com/D12268