Ensure that we use INSERT instead of LOAD DATA when inserting DataFrames with no columns
Summary: If we try to insert a DataFrame with an empty schema into, for instance, a table with columns that all have default values, we'll get a SQL syntax error because we try to run `LOAD DATA LOCAL INFILE ... INTO TABLE table_name()`, which is not valid.  We should always use `INSERT INTO table_name () VALUES ()` in this case.

Test Plan: manual

Reviewers: carl, ankur, joseph, hurshal

Reviewed By: hurshal

Subscribers: engineering-list

Differential Revision: http://grizzly.memsql.com/D12464