fixing T18611
Summary:
I caused this bug during the great rewrite when I changed the MemSQL Relation to
build it's own output rather than using the corresponding LogicalRelation's output.
Instead of changing back to what we had before - I had to grab the LogicalRelation's
output during pushdown and pass that into the BaseQuery instead.  The reason is that
when the DefaultSource is constructing a MemSQLRelation - we don't have access or
control over the wrapping LogicalRelation.

fixes T18611

Test Plan: manual

Reviewers: ankur, wsong, hurshal, john

Reviewed By: john

Subscribers: engineering-list

Maniphest Tasks: T18611

Differential Revision: http://grizzly.memsql.com/D13355