Spark 2.0 Connector: populate databaseName attribute of MemSQLQueryRelation to enable partition pushdown for custom-query loading
Summary: Currently we do not set a databaseName when the user provides a SQL query to load data from, since the query should be runnable by itself. However, the partition pushdown to the leaves requires a databaseName. In this diff we add a new parameter option for the user to specify the database name. If it is not provided, we grab the default database from the Spark config options.

Test Plan: Added tests to UserQuerySpec that analyze the created MemSQLQueryRelation to check that databaseName is not empty.

Reviewers: hurshal, john, carl

Reviewed By: carl

Subscribers: engineering-list

JIRA Issues: AP-1264

Differential Revision: https://grizzly.memsql.com/D18930