Update the connector to pull MemSQL connection settings from sparkSession.conf instead of the SparkContext
Summary:
The design doc here:
https://issues.apache.org/jira/browse/SPARK-13520

implies that sparkSession.conf is the correct place to look for MemSQL connection settings now as it may be thought of as a runtime configuration. The other configuration (which we currently use) accessed from the SparkContext is a "SparkConf", "immutable one Spark starts". These two configurations are the same at initialization, but users may want to connect to a different MemSQL cluster in the middle of a session. This change also makes using the connector with various notebooks easier where the SparkContext is already created.

Test Plan: Added ChangeSparkSessionConfSpec

Reviewers: hurshal, carl

Reviewed By: carl

Subscribers: engineering-list

JIRA Issues: AP-1317

Differential Revision: https://grizzly.memsql.com/D19813