Add option to enable jdbc connector to stream result sets row-by-row
Summary:
Currently we use the MySQL JDBC driver to connect to MemSQL. According to documentation,

By default, ResultSets are completely retrieved and stored in memory. In most cases this is the most efficient way to operate and, due to the design of the MySQL network protocol, is easier to implement. If you are working with ResultSets that have a large number of rows or large values and cannot allocate heap space in your JVM for the memory required, you can tell the driver to stream the results back one row at a time.

This diff adds an option to enable the one-at-a-time streaming on the driver. This can mitigate OOM errors for some types of work under limited executor memory, such as reading from MemSQL and writing the result to a csv file.

Test Plan: manual test

Reviewers: john, carl, hurshal

Reviewed By: hurshal

Subscribers: engineering-list

Differential Revision: https://grizzly.memsql.com/D21228