Fixed bugs with decimal literals and plancache
Summary:
When we were reading decimal literals by the query that was added before to the plancache,
precision and scale of the resule were taken from the old query.
Because of it we did one more cast that wasn't pushed down.
Added cast for every decimal literal

**Docs impact**:

Test Plan: suite tests

Reviewers: carl, iblinov-ua

Reviewed By: carl

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D40697