Forwardport: fix of the bug in codegen of join conditions
Summary:
In the old code if the condition was None we returned None and if we failed to pushdown we returned Some(None)
As a result, when we failed to pushdown we generated query with empty condition and when condition is empty we stopped the pushdown.
This beghaviour is fixed.
Also added more tests for joins and added check that both relations has the same jdbc connection options before joining them
**Docs impact**:

Test Plan: suite tests

Reviewers: carl, iblinov-ua

Subscribers: engineering-list

JIRA Issues: PLAT-4391

Differential Revision: https://grizzly.internal.memcompute.com/D41444