Fixed deadlock during data loading
Summary:
If the error occured before thread started to read from the input stream, it was not closed.
As a result - other thread was trying to write data to the full stream infinitelly.
**Design doc/spec**:
**Docs impact**: none
**Preliminary Reviewer(s)**:
**Final Reviewer**:
Test Plan:
Manually tested by throwing exceptions in different places of the code that executes the statement.

Reviewers:
pmishchenko-ua, carl
CC:

Jira issues:
PLAT-6261