Fixed deadlock in error handling
Summary:
Previously, if error occured in the thread that executes LOAD DATA queries and
stream was full we got a deadlock.
Now in the case of error, the stream is closed and the actual error is retrieved in abort method.
**Design doc/spec**:
**Docs impact**: none
**Preliminary Reviewer(s)**:
**Final Reviewer**:

Test Plan: https://webapp.io/memsql/commits?query=repo%3Asinglestore-spark-connector+id%3A361

Reviewers: carl, pmishchenko-ua

Subscribers: engineering-list

JIRA Issues: PLAT-6261

Differential Revision: https://grizzly.internal.memcompute.com/D57540